pyproject.toml
[project]
name = "dynamous_ai_tutor"
version = "0.1.0"
description = "A baseline AI agent."
requires-python = ">=3.10"
dependencies = [
    "instructor>=1.3.3",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "typer>=0.19.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = [
    ".",
]
